[ai-assisted] refactor(data): StaticModels SqlQuery 패키지 의존 분리
Issue: #441

Why:
- SqlQuery 제거 전 template-service의 비 SQL helper 의존을 SqlQuery 구현 패키지에서 분리해야 합니다.

What:
- StaticModels helper를 studio.one.platform.data.freemarker 패키지로 이동했습니다.
- 기존 sqlquery.factory.impl.StaticModels는 source/binary compatibility shim으로 남겨 새 helper에 위임하도록 했습니다.
- DynamicSqlNode와 FreemarkerTemplateBuilder가 새 helper를 사용하도록 변경했습니다.
- migration inventory에 StaticModels 분리 상태를 기록했습니다.

Validation:
- ./gradlew :studio-platform-data:build :studio-application-modules:template-service:build (SUCCESS)
- git diff --check (SUCCESS)
- Subagent review: code-reviewer/verifier 반복 리뷰 결과 반영 후 no actionable findings
