+++ studio-platform-data/src/main/java/studio/one/platform/data/sqlquery/builder/xml/dynamic/DynamicSqlNode.java
@@ -26,7 +26,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 
-import studio.one.platform.data.sqlquery.factory.impl.StaticModels;
+import studio.one.platform.data.freemarker.StaticModels;
 import freemarker.ext.beans.BeansWrapper;
 import freemarker.ext.beans.BeansWrapperBuilder;
 import freemarker.template.Configuration;
